moved theme switching logic into it's own component
