+++ packages/react-app/src/components/index.js
@@ -14,3 +14,4 @@ export { default as Header } from "./Header";
 export { default as Timeline } from "./Timeline";
 export { default as GasGauge } from "./GasGauge";
 export { default as BytesStringInput } from "./BytesStringInput";
+export { default as ThemeSwitch } from "./ThemeSwitch";
